--- a/addons/mail/static/src/core/common/chat_hub.js
+++ b/addons/mail/static/src/core/common/chat_hub.js
@@ -33,10 +33,7 @@ export class ChatHub extends Component {
         this.onResize();
         useExternalListener(browser, "resize", this.onResize);
         useEffect(() => {
-            if (
-                this.chatHub.actuallyFolded.length &&
-                this.store.channels?.status === "not_fetched"
-            ) {
+            if (this.chatHub.folded.length && this.store.channels?.status === "not_fetched") {
                 this.store.channels.fetch();
             }
         });
@@ -64,7 +61,7 @@ export class ChatHub extends Component {
 
     get hiddenCounter() {
         let counter = 0;
-        for (const chatWindow of this.chatHub.actuallyHidden) {
+        for (const chatWindow of this.chatHub.folded.slice(this.chatHub.maxFolded)) {
             counter += chatWindow.thread.importantCounter > 0 ? 1 : 0;
         }
         return counter;
@@ -73,7 +70,7 @@ export class ChatHub extends Component {
     expand() {
         this.chatHub.compact = false;
         Object.assign(this.compactPosition, { left: "auto", top: "auto" });
-        this.more.isOpen = this.chatHub.actuallyHidden.length !== 0;
+        this.more.isOpen = this.chatHub.folded.length > this.chatHub.maxFolded;
     }
 }