--- a/addons/mail/static/src/core/common/chat_hub_model.js
+++ b/addons/mail/static/src/core/common/chat_hub_model.js
@@ -25,7 +25,6 @@ export class ChatHub extends Record {
         inverse: "hubAsOpened",
         /** @this {import("models").ChatHub} */
         onAdd(r) {
-            this.folded.delete(r);
             this.onRecompute();
         },
         /** @this {import("models").ChatHub} */
@@ -38,7 +37,6 @@ export class ChatHub extends Record {
         inverse: "hubAsFolded",
         /** @this {import("models").ChatHub} */
         onAdd(r) {
-            this.opened.delete(r);
             this.onRecompute();
         },
         /** @this {import("models").ChatHub} */
@@ -46,9 +44,6 @@ export class ChatHub extends Record {
             this.onRecompute();
         },
     });
-    actuallyOpened = Record.many("ChatWindow", { inverse: "hubAsActuallyOpened" });
-    actuallyFolded = Record.many("ChatWindow");
-    actuallyHidden = Record.many("ChatWindow");
 
     closeAll() {
         [...this.opened, ...this.folded].forEach((cw) => cw.close());
@@ -58,13 +53,6 @@ export class ChatHub extends Record {
         while (this.opened.length > this.maxOpened) {
             const cw = this.opened.pop();
             this.folded.unshift(cw);
-        }
-        this.actuallyOpened = this.opened;
-        this.actuallyFolded = this.folded;
-        this.actuallyHidden.clear();
-        while (this.actuallyFolded.length > this.maxFolded) {
-            const cw = this.actuallyFolded.pop();
-            this.actuallyHidden.unshift(cw);
         }
     }