--- a/addons/mail/static/src/core/common/chat_window.js
+++ b/addons/mail/static/src/core/common/chat_window.js
@@ -103,13 +103,11 @@ export class ChatWindow extends Component {
                 this.close({ escape: true });
                 break;
             case "Tab": {
-                const index = this.store.chatHub.actuallyOpened.findIndex((cw) =>
-                    cw.eq(chatWindow)
-                );
-                if (index === this.store.chatHub.actuallyOpened.length - 1) {
-                    this.store.chatHub.actuallyOpened[0].focus();
+                const index = this.store.chatHub.opened.findIndex((cw) => cw.eq(chatWindow));
+                if (index === this.store.chatHub.opened.length - 1) {
+                    this.store.chatHub.opened[0].focus();
                 } else {
-                    this.store.chatHub.actuallyOpened[index + 1].focus();
+                    this.store.chatHub.opened[index + 1].focus();
                 }
                 break;
             }