@robodoo 2024-08-23
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/177736.png)](https://mergebot.odoo.com/odoo/odoo/pull/177736)